chore: retire the v0.6 send path and the self-funded VerifyingPaymaster (#721)
