+++ config/worker-base-sepolia.example.yaml
@@ -24,6 +24,10 @@ health_address: "0.0.0.0:8091"
 gateway_address: "127.0.0.1:2206"
 
 # Chain RPC endpoints (Base Sepolia). Set via .env.local / env.
+# Supply a chain RPC you control. Do NOT point this at a free public endpoint
+# (publicnode, *.base.org, ankr, llamarpc, ...): they rate-limit under load,
+# which makes runs fail nondeterministically on whichever request happens to be
+# in flight rather than on anything real.
 eth_rpc_url: ${BASE_SEPOLIA_RPC_URL}
 eth_ws_url:  ${BASE_SEPOLIA_WS_URL}
