+++ pkg/erc4337/preset/sent_userop.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/ethclient"
 
 	"github.com/AvaProtocol/EigenLayer-AVS/core/config"
 	"github.com/AvaProtocol/EigenLayer-AVS/pkg/erc4337/userop"
@@ -44,91 +43,51 @@ type SentUserOp struct {
 	Sponsored bool
 }
 
-// SendUserOpAuto sends an operation using whichever account implementation the
-// chain is configured for, and reports the result in version-neutral terms.
+// SendUserOpAuto sends an operation and reports the result in version-neutral
+// terms.
 //
-// Dispatch is per CHAIN, not per wallet. That is a deliberate consequence of
-// cutting every chain over at once: a wallet created before the cutover is a
-// v0.6 SimpleAccount, and on a chain now configured for MA v2 it will no
-// longer execute. Those wallets were audited to hold no meaningful balances
-// before this was chosen. If that assumption ever stops holding, this is the
-// function that has to learn to look at the wallet's factory instead.
+// It is the single send entry point. The v0.6 path it used to dispatch to is
+// gone: every chain runs Modular Account v2 on EntryPoint v0.7, and a config
+// asking for anything else is refused at load (see SmartWalletConfig
+// validation) rather than routed to an EntryPoint nothing deploys against.
 //
-// paymasterReq and executionFeeWei apply only to the v0.6 path. MA v2
-// sponsorship comes from the Gas Manager policy on the chain's config, so
-// those arguments are ignored there rather than quietly changing meaning.
+// Sponsorship comes from the chain's Alchemy Gas Manager policy. The
+// paymasterReq and executionFeeWei arguments this used to take applied only to
+// the v0.6 verifying paymaster and were already ignored here, so they are gone
+// rather than left as parameters that quietly mean nothing.
 func SendUserOpAuto(
 	smartWalletConfig *config.SmartWalletConfig,
 	owner common.Address,
 	callData []byte,
-	paymasterReq *VerifyingPaymasterRequest,
 	senderOverride *common.Address,
 	saltOverride *big.Int,
-	executionFeeWei *big.Int,
 	lgr logger.Logger,
 ) (*SentUserOp, *types.Receipt, error) {
 	if smartWalletConfig == nil {
 		return nil, nil, fmt.Errorf("nil smart wallet config")
 	}
-
-	if smartWalletConfig.UsesModularAccountV2() {
-		// nil auth: SendUserOpMAv2 resolves the session grant itself, AFTER
-		// it has resolved the actual sender. Grants are stored per
-		// smart-wallet address, and callers do not always pass an override —
-		// resolving here against the owner EOA used to find nothing and sign
-		// as the fallback entity the gateway cannot validate for.
-		op, receipt, err := SendUserOpMAv2(smartWalletConfig, owner, callData, senderOverride, saltOverride, nil, lgr)
-		sent, convErr := sentFromV07(op, big.NewInt(smartWalletConfig.ChainID))
-		if err != nil {
-			// Report the send failure, not the conversion: the send is why
-			// this failed, and `sent` is best-effort context for the caller.
-			return sent, receipt, err
-		}
-		if convErr != nil {
-			return nil, receipt, convErr
-		}
-		return sent, receipt, nil
+	if !smartWalletConfig.UsesModularAccountV2() {
+		return nil, nil, fmt.Errorf(
+			"chain %d is configured for account provider %q; only %s can execute — "+
+				"the v0.6 send path was removed with the EntryPoint v0.7 cutover",
+			smartWalletConfig.ChainID, smartWalletConfig.AccountProviderName(),
+			config.AccountProviderModularAccountV2)
 	}
 
-	op, receipt, err := SendUserOp(smartWalletConfig, owner, callData, paymasterReq,
-		senderOverride, saltOverride, executionFeeWei, lgr)
-	sent := sentFromV06(op, smartWalletConfig.EntrypointAddress, big.NewInt(smartWalletConfig.ChainID))
+	// nil auth: SendUserOpMAv2 resolves the session grant itself, AFTER it has
+	// resolved the actual sender. Grants are stored per smart-wallet address,
+	// and callers do not always pass an override — resolving here against the
+	// owner EOA used to find nothing and sign as the fallback entity the
+	// gateway cannot validate for.
+	op, receipt, err := SendUserOpMAv2(smartWalletConfig, owner, callData, senderOverride, saltOverride, nil, lgr)
+	sent, convErr := sentFromV07(op, big.NewInt(smartWalletConfig.ChainID))
 	if err != nil {
+		// Report the send failure, not the conversion: the send is why this
+		// failed, and `sent` is best-effort context for the caller.
 		return sent, receipt, err
 	}
-	return sent, receipt, nil
-}
-
-// SendUserOpAutoWithWsClient is SendUserOpAuto reusing a caller-owned
-// WebSocket client for receipt monitoring.
-//
-// The MA v2 path ignores wsClient and opens its own from the chain config.
-// Sharing the aggregator's long-lived socket saved a dial on the v0.6 path;
-// threading it through the v0.7 builder is not worth a second signature for
-// what is only a receipt watcher.
-func SendUserOpAutoWithWsClient(
-	smartWalletConfig *config.SmartWalletConfig,
-	owner common.Address,
-	callData []byte,
-	paymasterReq *VerifyingPaymasterRequest,
-	senderOverride *common.Address,
-	saltOverride *big.Int,
-	wsClient *ethclient.Client,
-	executionFeeWei *big.Int,
-	lgr logger.Logger,
-) (*SentUserOp, *types.Receipt, error) {
-	if smartWalletConfig == nil {
-		return nil, nil, fmt.Errorf("nil smart wallet config")
-	}
-	if smartWalletConfig.UsesModularAccountV2() {
-		return SendUserOpAuto(smartWalletConfig, owner, callData, paymasterReq,
-			senderOverride, saltOverride, executionFeeWei, lgr)
-	}
-	op, receipt, err := SendUserOpWithWsClient(smartWalletConfig, owner, callData, paymasterReq,
-		senderOverride, saltOverride, wsClient, executionFeeWei, lgr)
-	sent := sentFromV06(op, smartWalletConfig.EntrypointAddress, big.NewInt(smartWalletConfig.ChainID))
-	if err != nil {
-		return sent, receipt, err
+	if convErr != nil {
+		return nil, receipt, convErr
 	}
 	return sent, receipt, nil
 }
@@ -158,21 +117,4 @@ func sentFromV07(op *userop.UserOperationV07, chainID *big.Int) (*SentUserOp, er
 	}
 	sent.UserOpHash = hash
 	return sent, nil
-}
-
-func sentFromV06(op *userop.UserOperation, entryPoint common.Address, chainID *big.Int) *SentUserOp {
-	if op == nil {
-		return nil
-	}
-	return &SentUserOp{
-		Sender:               op.Sender,
-		Nonce:                op.Nonce,
-		EntryPoint:           entryPoint,
-		UserOpHash:           op.GetUserOpHash(entryPoint, chainID),
-		CallGasLimit:         op.CallGasLimit,
-		VerificationGasLimit: op.VerificationGasLimit,
-		PreVerificationGas:   op.PreVerificationGas,
-		MaxFeePerGas:         op.MaxFeePerGas,
-		Sponsored:            len(op.PaymasterAndData) > 0,
-	}
 }
